Modified existing convention for referencing Dlls from bin/obj to include dotnet sdk (.net core) and reference assemblies. Tested on self.
