+++ Conventional.Tests/DogFoodConventions.cs
@@ -22,7 +22,15 @@ namespace Conventional.Tests
                 .WithFailureAssertion(Assert.Fail);
         }
 
-        private Type[] ConventionTypes => new[]
+        [Test]
+        public void GivenAPattern_LocatesAndReturnsAllAssembliesForThatPattern()
+        {
+            AllAssemblies.WithNamesMatching("*")
+                .MustConformTo(Convention.MustNotReferenceDllsFromTransientOrSdkDirectories)
+                .WithFailureAssertion(Assert.Fail);
+        }
+
+        private static Type[] ConventionTypes => new[]
         {
             typeof (ConventionSpecification),
             typeof (AssemblyConventionSpecification),
